servo: Merge #5441 - Use usize rather than u64 in MemoryReport (from servo:heap-unclassified); r=jdm
Source-Repo: https://github.com/servo/servo
Source-Revision: 8976bbfe7d8fa061ee7270e4b3553f7fde6050e6

UltraBlame original commit: 0dda25a8d154a47f2913fb4c7ac370c700d9482f

+++ servo/components/profile/mem.rs
@@ -32,12 +32,13 @@ macro_rules! path {
     }}
 }
 
+
 pub struct Report {
     
     pub path: Vec<String>,
 
     
-    pub size: u64,
+    pub size: usize,
 }
 
 
@@ -208,7 +209,7 @@ impl Profiler {
 struct ReportsTree {
     
     
-    size: u64,
+    size: usize,
 
     
     
@@ -243,7 +244,7 @@ impl ReportsTree {
     }
 
     
-    fn insert(&mut self, path: &[String], size: u64) {
+    fn insert(&mut self, path: &[String], size: usize) {
         let mut t: &mut ReportsTree = self;
         for path_seg in path.iter() {
             let i = match t.find_child(&path_seg) {
@@ -264,7 +265,7 @@ impl ReportsTree {
 
     
     
-    fn compute_interior_node_sizes(&mut self) -> u64 {
+    fn compute_interior_node_sizes(&mut self) -> usize {
         if !self.children.is_empty() {
             
             if self.size != 0 {
@@ -313,7 +314,7 @@ impl ReportsForest {
     }
 
     
-    fn insert(&mut self, path: &[String], size: u64) {
+    fn insert(&mut self, path: &[String], size: usize) {
         
         if !self.trees.contains_key(&path[0]) {
             self.trees.insert(path[0].clone(), ReportsTree::new(path[0].clone()));
@@ -440,7 +441,7 @@ mod system_reporter {
     }
 
     #[cfg(target_os="linux")]
-    fn get_system_heap_allocated() -> Option<u64> {
+    fn get_system_heap_allocated() -> Option<usize> {
         let mut info: struct_mallinfo;
         unsafe {
             info = mallinfo();
@@ -449,11 +450,11 @@ mod system_reporter {
         
         
         
-        Some((info.hblkhd + info.uordblks) as u64)
+        Some((info.hblkhd + info.uordblks) as usize)
     }
 
     #[cfg(not(target_os="linux"))]
-    fn get_system_heap_allocated() -> Option<u64> {
+    fn get_system_heap_allocated() -> Option<usize> {
         None
     }
 
@@ -462,7 +463,7 @@ mod system_reporter {
                       newp: *mut c_void, newlen: size_t) -> c_int;
     }
 
-    fn get_jemalloc_stat(value_name: &str) -> Option<u64> {
+    fn get_jemalloc_stat(value_name: &str) -> Option<usize> {
         
         
         
@@ -494,7 +495,7 @@ mod system_reporter {
             return None;
         }
 
-        Some(value as u64)
+        Some(value as usize)
     }
 
     
@@ -503,7 +504,7 @@ mod system_reporter {
     );
 
     #[cfg(target_os="linux")]
-    fn get_proc_self_statm_field(field: usize) -> Option<u64> {
+    fn get_proc_self_statm_field(field: usize) -> Option<usize> {
         use std::fs::File;
         use std::io::Read;
 
@@ -511,42 +512,42 @@ mod system_reporter {
         let mut contents = String::new();
         option_try!(f.read_to_string(&mut contents).ok());
         let s = option_try!(contents.words().nth(field));
-        let npages = option_try!(s.parse::<u64>().ok());
-        Some(npages * (::std::env::page_size() as u64))
+        let npages = option_try!(s.parse::<usize>().ok());
+        Some(npages * ::std::env::page_size())
     }
 
     #[cfg(target_os="linux")]
-    fn get_vsize() -> Option<u64> {
+    fn get_vsize() -> Option<usize> {
         get_proc_self_statm_field(0)
     }
 
     #[cfg(target_os="linux")]
-    fn get_resident() -> Option<u64> {
+    fn get_resident() -> Option<usize> {
         get_proc_self_statm_field(1)
     }
 
     #[cfg(target_os="macos")]
-    fn get_vsize() -> Option<u64> {
+    fn get_vsize() -> Option<usize> {
         virtual_size()
     }
 
     #[cfg(target_os="macos")]
-    fn get_resident() -> Option<u64> {
+    fn get_resident() -> Option<usize> {
         resident_size()
     }
 
     #[cfg(not(any(target_os="linux", target_os = "macos")))]
-    fn get_vsize() -> Option<u64> {
+    fn get_vsize() -> Option<usize> {
         None
     }
 
     #[cfg(not(any(target_os="linux", target_os = "macos")))]
-    fn get_resident() -> Option<u64> {
+    fn get_resident() -> Option<usize> {
         None
     }
 
     #[cfg(target_os="linux")]
-    fn get_resident_segments() -> Vec<(String, u64)> {
+    fn get_resident_segments() -> Vec<(String, usize)> {
         use regex::Regex;
         use std::collections::HashMap;
         use std::collections::hash_map::Entry;
@@ -575,7 +576,7 @@ mod system_reporter {
         let rss_re = Regex::new(r"^Rss: +(\d+) kB").unwrap();
 
         
-        let mut seg_map: HashMap<String, u64> = HashMap::new();
+        let mut seg_map: HashMap<String, usize> = HashMap::new();
 
         #[derive(PartialEq)]
         enum LookingFor { Segment, Rss }
@@ -620,7 +621,7 @@ mod system_reporter {
                     Some(cap) => cap,
                     None => continue,
                 };
-                let rss = cap.at(1).unwrap().parse::<u64>().unwrap() * 1024;
+                let rss = cap.at(1).unwrap().parse::<usize>().unwrap() * 1024;
 
                 if rss > 0 {
                     
@@ -639,7 +640,7 @@ mod system_reporter {
             }
         }
 
-        let mut segs: Vec<(String, u64)> = seg_map.into_iter().collect();
+        let mut segs: Vec<(String, usize)> = seg_map.into_iter().collect();
 
         
         
@@ -650,7 +651,7 @@ mod system_reporter {
     }
 
     #[cfg(not(target_os="linux"))]
-    fn get_resident_segments() -> Vec<(String, u64)> {
+    fn get_resident_segments() -> Vec<(String, usize)> {
         vec![]
     }
 }
